+++ CHANGELOG.md
@@ -5,3 +5,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### v0.1.0
 
 * Initial release.
+* Add cop Platanus/NoCommand